added some spacings and borders
I had to use last-of-type instead of last-child, because the last child
was a link to the archive. I also realized I can use the border
attribute without having to specify colors. At the moment, there is no
guarantee that this layout is in rhythm, especially because the images
aren't resized yet.