concentrated mostly on forms and news
Decided to rename $header/$body to $serif/$sans-serif, as I wasn't using
$header only for headers and vice-versa. I would actually be suprised if
this [type] stunt with input elements works in those browsers that
support it. As #login can now be two elements, I had to be more
specific to avoid bugs. I converted some px to em, but it's still not
finished. I removed that addition of colon in #facebook, because I
decided to do all these things in JS, it's more supported AFAIK. I'm
still in the process of removing h2, that's why I hid it.