fixed a lot of bugs and issues
I eliminated a lot of code and implemented it as mixins. I styled login
and post forms, so they're a lot prettier to look at now. I used 91%
because 100% extends the width beyond the width of the td element. I
don't know if it's only a Chrome issue, but I'm pretty sure it's not,
even if it doesn't make a lot of sense. I used the way of styling the
add button to style other buttons, which aren't in .controls. Lastly, I
used some math to automatize the #facebook childrens' widths a bit.