Style the admin member list
In order for my CSS to be able to recognize the difference between
user's and admin's member list, I had to do some pontentially confusing
styling.

I commented out the form for adding a member because I think think it
should appear after admin clicks on the add button. I also commented out
the error report because there was an error on execution. I commented
out the finish button because I don't think we'll have one. This was
obviously temporary hiding.