Fix and style buttons for adding/removing pages
I decided to use the "subtle" mixin on them, but I had to tweak it a
little so it can cancel the effect of the "button" mixin. The addition
button was showing even when the admin wasn't logged in, so I fix that.
I had to fix the JS part because the selector was also targeting
the addition button in the navigation. I added the #pages class to the
page links in the navigation, because simple "ol" selector was targeting
the removal button. I also swapped addition and removal buttons.